# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.AddItemInput:
    properties:
      item_id:
        type: string
      user_review:
        type: string
    required:
    - item_id
    - user_review
    type: object
...
  handler.CollectionItemResponse:
    properties:
      cover_image:
        type: string
      creator_id:
        type: string
      description:
        type: string
      id:
        type: string
      is_custom:
        type: boolean
      is_public:
        type: boolean
      title:
        type: string
      type:
        type: string
    type: object
...
  handler.CreateCollectionItemInput:
    properties:
      cover_image:
        type: string
      description:
        type: string
      is_custom:
        type: boolean
      is_public:
        type: boolean
      title:
        type: string
      type:
        type: string
    required:
    - title
    - type
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.IDResponse:
    properties:
      id:
        type: string
    type: object
...
  handler.PaginatedItemsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.CollectionItemResponse'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handler.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Memoria API
  version: "1.0"
paths:
...
  /collections/{id}/items:
    get:
      description: Возвращает все элементы указанной коллекции
      parameters:
      - description: ID коллекции
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список элементов коллекции
          schema: {}
        "400":
          description: Не указан ID коллекции
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Элементы не найдены
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить элементы коллекции
      tags:
      - collections
    post:
      consumes:
      - application/json
      description: Добавляет существующий элемент в указанную коллекцию
      parameters:
      - description: ID коллекции
        in: path
        name: id
        required: true
        type: string
      - description: Данные для добавления элемента
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.AddItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: Элемент успешно добавлен
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Коллекция или элемент не найдены
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Элемент уже в коллекции
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Добавить элемент в коллекцию
      tags:
      - collections
  /items:
    get:
      description: Возвращает пагинированный список элементов по указанному типу
      parameters:
      - default: book
        description: Тип элементов
        in: query
        name: type
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Кол-во элементов на странице
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handler.PaginatedItemsResponse'
        "502":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить элементы по типу
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Создает новый элемент коллекции
      parameters:
      - description: Данные для создания элемента
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCollectionItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: Элемент создан, возвращает ID
          schema:
            $ref: '#/definitions/handler.IDResponse'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создать элемент коллекции
      tags:
      - items
...
securityDefinitions:
  apiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"